requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
gunicorn
pandas